web: move some utility functions to a util module
Mostly template filter functions at the moment.
