+++ web/src/html/user.rs
@@ -1,8 +1,8 @@
 use crate::{
-    app_url,
     auth::SqliteUser,
     error::{self, Error},
     state::AppState,
+    util::app_url,
     Message, MessageType, TemplateKey,
 };
 use anyhow::{anyhow, Context};
