+++ web/src/util.rs
@@ -1,7 +1,78 @@
+use std::collections::HashMap;
+
+use axum::http::Uri;
+use minijinja::ErrorKind;
 use pulldown_cmark::Event;
 use regex::Regex;
 
-use crate::app_url;
+use crate::APP_PREFIX;
+
+pub fn app_url(value: &str) -> String {
+    [APP_PREFIX, value.trim_start_matches('/')].join("")
+}
+
+pub fn append_query_param(
+    uristr: String,
+    key: String,
+    value: String,
+) -> Result<String, minijinja::Error> {
+    let uri = uristr.parse::<Uri>().map_err(|e| {
+        minijinja::Error::new(ErrorKind::InvalidOperation, "Unable to parse uri string")
+            .with_source(e)
+    })?;
+    let mut query: HashMap<_, _> = match uri.query() {
+        Some(q) => serde_urlencoded::from_str(q).map_err(|e| {
+            minijinja::Error::new(ErrorKind::InvalidOperation, "Unable to decode query params")
+                .with_source(e)
+        })?,
+        None => HashMap::new(),
+    };
+    query.insert(key.as_str(), value.as_str());
+    let querystring = serde_urlencoded::to_string(query).map_err(|e| {
+        minijinja::Error::new(ErrorKind::InvalidOperation, "Unable to encode query params")
+            .with_source(e)
+    })?;
+
+    Ok(format!("?{querystring}"))
+}
+
+pub fn truncate_text(mut s: String, chars: Option<usize>) -> String {
+    let chars = chars.unwrap_or(100);
+    if s.len() > chars {
+        s.truncate(chars);
+        s + "..."
+    } else {
+        s
+    }
+}
+
+pub fn format_id_number(id: i64, prefix: Option<&str>, width: Option<usize>) -> String {
+    let width = width.unwrap_or(4);
+    let prefix = prefix.unwrap_or("");
+    format!("{}{:0>width$}", prefix, id, width = width)
+}
+
+pub fn format_quantity(qty: f64) -> String {
+    let mut metric_qty = qty;
+    let mut metric_label = "grams";
+    let imperial_qty = metric_qty * 0.03527396195;
+
+    if metric_qty > 1000.0 {
+        metric_label = "kilograms";
+        metric_qty /= 1000.0;
+    }
+    let metric = format!("{metric_qty:.2} {metric_label}");
+
+    let imperial = if imperial_qty > 16.0 {
+        let lbs = (imperial_qty / 16.0).floor() as i64;
+        let oz = imperial_qty % 16.0;
+        format!("{lbs} lbs {oz:.2} ounces")
+    } else {
+        format!("{imperial_qty:.2} ounces")
+    };
+
+    format!("{metric} ({imperial})")
+}
 
 pub fn markdown(value: Option<&str>) -> minijinja::Value {
     let value = value.unwrap_or("");
